refactor(cli): move the convert group into its own module
Cuts `gpio convert` out of cli/main.py into
geoparquet_io/cli/commands/convert.py, unchanged, along with the two helpers
only it uses -- `_convert_streaming` and `_reproject_impl_cli` -- and the
`ConvertDefaultGroup` class that drives its extension-based dispatch. The
class keeps its name, which the CLI surface snapshot pins.

`tests/test_cli_api_call_parity_scaffold.py` resolved `convert_to_geoparquet`
through `geoparquet_io.cli.main` and now passes `module=cli_convert`, the
mechanism the scaffold already documents for extracted groups.

tests/data/cli_surface.json is byte-identical.
